split chilldkg into module files and move tests
replaces 003399dda912d5e17a22e29884622a90dff56e8d
